YOUTUBE_CHATBOT/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.57.1
anyio==3.7.1
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.0
base58==2.1.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
brotlipy==0.7.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
comtypes==1.4.8
contourpy==1.3.0
cryptography==44.0.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.15
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.116.1
filelock==3.12.4
filetype==1.2.0
Flask==2.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-PyMongo==2.3.0
flatbuffers==25.2.10
fonttools==4.55.0
frozenlist==1.7.0
fsspec==2023.9.1
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
h2==3.2.0
h5py==3.14.0
hpack==3.0.0
hstspreload==2024.12.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
imutils==0.5.4
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.0.1
jedi==0.19.2
Jinja2==3.0.1
jiter==0.8.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.10.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
langchain==0.3.26
langchain-anthropic==0.3.17
langchain-community==0.3.27
langchain-core==0.3.68
langchain-experimental==0.3.4
langchain-google-genai==2.0.10
langchain-huggingface==0.3.0
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langgraph==0.5.3
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.73
langsmith==0.4.4
lazy_loader==0.3
libclang==18.1.1
librosa==0.10.1
llvmlite==0.40.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.26.1
matplotlib==3.9.4
matplotlib-inline==0.1.2
mdurl==0.1.2
ml_dtypes==0.5.1
mne==1.8.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.6.3
mypy_extensions==1.1.0
namex==0.0.8
narwhals==1.46.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numba==0.57.1
numpy==2.0.2
oauthlib==3.2.2
onnx==1.10.1
openai==1.93.1
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.13.1
orjson==3.10.18
ormsgpack==1.10.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.1
parso==0.8.4
Pillow==9.4.0
platformdirs==3.10.0
playsound==1.3.0
pluggy==1.5.0
pooch==1.7.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==3.20.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.10.1
pydantic_core==2.27.1
pydeck==0.9.1
pygame==2.6.1
Pygments==2.18.0
pymongo==4.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.0
pypdf==5.7.0
pypiwin32==223
pyserial==3.5
pysound==0.2
pytest==8.3.4
python-dateutil==2.8.2
python-dotenv==1.1.1
pyttsx3==2.98
pytube==15.0.0
pytz==2023.3.post1
pywin32==308
PyYAML==6.0.1
pyzmq==27.0.0
referencing==0.35.1
regex==2023.8.8
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
selenium==4.0.0b4
sentence-transformers==5.0.0
sentencepiece==0.2.0
six==1.15.0
sklearn==0.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.3.6
SpeechRecognition==3.11.0
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.47.2
streamlit==1.46.1
sympy==1.14.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.0
tensorflow==2.19.0
tensorflow-estimator==2.5.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==1.1.0
tf2onnx==1.9.3
tf_keras==2.18.0
threadpoolctl==3.2.0
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
torch==2.7.1+cpu
torch-tb-profiler==0.4.3
torchaudio==2.7.1+cpu
torchvision==0.22.1+cpu
tornado==6.4.1
tqdm==4.66.1
traitlets==5.14.3
transformers==4.53.2
trio==0.29.0
trio-websocket==0.12.2
typing==3.7.4.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2023.3
tzlocal==5.2
uritemplate==4.2.0
urllib3==1.26.20
urllib3-secure-extra==0.1.0
uvicorn==0.35.0
validators==0.34.0
virtualenv==20.27.1
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==2.1.2
wrapt==1.12.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
youtube-transcript-api==1.1.1
yt-dlp==2024.12.13
zipp==3.21.0
zstandard==0.23.0